requirements.txt
google-generativeai>=0.8.0
pandas>=2.0.0
numpy<2.0.0
matplotlib>=3.6.0,<3.8.0
seaborn>=0.12.0
plotly>=5.15.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
scikit-learn>=1.3.0
scipy>=1.10.0
typing-extensions>=4.5.0
python-multipart>=0.0.6
asyncio>=3.4.3
